test(AglaError): restructure and simplify unit tests
- Removed redundant and obvious tests from AglaError specs
- Split and reorganized tests into focused files:
  - core (constructor, properties, legacy context, edge cases)
  - chaining (message concatenation, context merge, error causes)
  - serialization (JSON, toString, circular refs, full property set)
  - inheritance (Error extension, name, stack trace, instanceof)
- Simplified ErrorSeverity tests by removing redundant enum checks while keeping edge cases
- Removed unnecessary TypeScript compatibility tests and added runtime type guard tests
- Deleted obsolete specs and test-only types (HttpHeaders, TAglaErrorContextWithSymbols, etc.)
- Moved ErrorSeverity tests into `__tests__/unit/`
