--- shared/packages/types/src/__tests__/HttpHeaders.validation.spec.ts
@@ -1,82 +1,0 @@
-// src: src/__tests__/HttpHeaders.validation.spec.ts
-// @(#) : HttpHeaders型バリデーション機能のテスト
-//
-// Copyright (c) 2025 atsushifx <http://github.com/atsushifx>
-//
-// This software is released under the MIT License.
-// https://opensource.org/licenses/MIT
-
-// Testing framework
-import { describe, expect, it } from 'vitest';
-
-// Type definitions
-import type { _THttpHeaders } from 'src/__tests__/helpers/test-types.types';
-
-/**
- * HttpHeaders Type Definition Tests
- * Tests type compatibility for HttpHeaders type (R-002-01)
- */
-describe('Given HttpHeaders type definition', () => {
-  describe('When creating header objects', () => {
-    it('Then should satisfy HttpHeaders type compatibility', () => {
-      // Arrange
-      const headers: _THttpHeaders = { 'content-type': 'application/json' };
-
-      // Act & Assert - Type compatibility test as per R-002-01
-      expect(headers).toSatisfy((h): h is _THttpHeaders => {
-        return typeof h === 'object' && h !== null
-          && Object.values(h).every((val) => typeof val === 'string');
-      });
-    });
-
-    it('Then should handle multiple header values', () => {
-      // Arrange
-      const multipleHeaders: _THttpHeaders = {
-        'content-type': 'application/json',
-        'authorization': 'Bearer token123',
-        'x-custom-header': 'custom-value',
-      };
-
-      // Act & Assert
-      expect(multipleHeaders).toSatisfy((h): h is _THttpHeaders => {
-        return typeof h === 'object' && h !== null
-          && Object.values(h).every((val) => typeof val === 'string');
-      });
-    });
-
-    it('Then should handle empty headers object', () => {
-      // Arrange
-      const emptyHeaders: _THttpHeaders = {};
-
-      // Act & Assert
-      expect(emptyHeaders).toSatisfy((h): h is _THttpHeaders => {
-        return typeof h === 'object' && h !== null;
-      });
-    });
-  });
-
-  describe('When validating header structure', () => {
-    it('Then should enforce string values only', () => {
-      // Arrange - This should cause TypeScript compilation error if uncommented
-      // const invalidHeaders: HttpHeaders = { 'content-type': 123 };
-
-      // Act & Assert - Verify type constraint through successful compilation
-      const validHeaders: _THttpHeaders = { 'content-type': 'application/json' };
-      expect(typeof validHeaders['content-type']).toBe('string');
-    });
-
-    it('Then should allow arbitrary header names', () => {
-      // Arrange
-      const customHeaders: _THttpHeaders = {
-        'x-custom-1': 'value1',
-        'x-custom-2': 'value2',
-        'Authorization': 'Bearer xyz',
-        'content-type': 'text/html',
-      };
-
-      // Act & Assert
-      expect(Object.keys(customHeaders)).toHaveLength(4);
-      expect(Object.values(customHeaders).every((val) => typeof val === 'string')).toBe(true);
-    });
-  });
-});
