--- shared/packages/types/src/__tests__/TTestBuffer.validation.spec.ts
@@ -1,163 +1,0 @@
-// src: src/__tests__/_TTestBuffer.validation.spec.ts
-// @(#) : _TTestBuffer型テスト専用型のテスト
-//
-// Copyright (c) 2025 atsushifx <http://github.com/atsushifx>
-//
-// This software is released under the MIT License.
-// https://opensource.org/licenses/MIT
-
-// Testing framework
-import { describe, expect, it } from 'vitest';
-
-// Type definitions
-import type { _TTestBuffer } from './helpers/test-types.types.js';
-
-/**
- * _TTestBuffer Test Type Definition Tests
- * Tests type compatibility for _TTestBuffer type (R-003-03)
- */
-describe('Given _TTestBuffer test type definition', () => {
-  describe('When creating test buffer objects', () => {
-    it('Then should satisfy _TTestBuffer type compatibility', () => {
-      // Arrange
-      const buffer: _TTestBuffer = { data: 'test', count: 42, active: true };
-
-      // Act & Assert - Type compatibility test as per R-003-03
-      expect(buffer).toSatisfy((b): b is _TTestBuffer => {
-        return typeof b === 'object' && b !== null;
-      });
-    });
-
-    it('Then should handle various data types as values', () => {
-      // Arrange
-      const mixedBuffer: _TTestBuffer = {
-        stringValue: 'text',
-        numberValue: 123,
-        booleanValue: true,
-        arrayValue: [1, 2, 3],
-        objectValue: { nested: 'data' },
-        nullValue: null,
-        undefinedValue: undefined,
-        dateValue: new Date(),
-        functionValue: () => 'test',
-      };
-
-      // Act & Assert
-      expect(mixedBuffer).toSatisfy((b): b is _TTestBuffer => {
-        return typeof b === 'object' && b !== null;
-      });
-      expect(typeof mixedBuffer.stringValue).toBe('string');
-      expect(typeof mixedBuffer.numberValue).toBe('number');
-      expect(typeof mixedBuffer.booleanValue).toBe('boolean');
-      expect(Array.isArray(mixedBuffer.arrayValue)).toBe(true);
-      expect(typeof mixedBuffer.objectValue).toBe('object');
-      expect(mixedBuffer.nullValue).toBeNull();
-      expect(mixedBuffer.undefinedValue).toBeUndefined();
-    });
-
-    it('Then should handle empty test buffer', () => {
-      // Arrange
-      const emptyBuffer: _TTestBuffer = {};
-
-      // Act & Assert
-      expect(emptyBuffer).toSatisfy((b): b is _TTestBuffer => {
-        return typeof b === 'object' && b !== null;
-      });
-      expect(Object.keys(emptyBuffer)).toHaveLength(0);
-    });
-
-    it('Then should support dynamic property addition', () => {
-      // Arrange
-      const dynamicBuffer: _TTestBuffer = {};
-
-      // Act - Add properties dynamically
-      dynamicBuffer.dynamicProp1 = 'added at runtime';
-      dynamicBuffer.dynamicProp2 = 456;
-      dynamicBuffer.dynamicProp3 = { added: true };
-
-      // Assert
-      expect(dynamicBuffer.dynamicProp1).toBe('added at runtime');
-      expect(dynamicBuffer.dynamicProp2).toBe(456);
-      expect(dynamicBuffer.dynamicProp3).toEqual({ added: true });
-      expect(Object.keys(dynamicBuffer)).toHaveLength(3);
-    });
-  });
-
-  describe('When validating buffer structure', () => {
-    it('Then should allow any unknown value types', () => {
-      // Arrange - Test various complex unknown types
-      const complexBuffer: _TTestBuffer = {
-        simpleString: 'value',
-        complexObject: {
-          deep: {
-            nested: {
-              structure: 'works',
-            },
-          },
-        },
-        genericArray: ['string', 123, true, null, { mixed: 'array' }],
-      };
-
-      // Act & Assert - Verify complex structures work
-      expect(complexBuffer.simpleString).toBe('value');
-      expect((complexBuffer.complexObject as { deep: { nested: { structure: string } } }).deep.nested.structure).toBe(
-        'works',
-      );
-      expect((complexBuffer.genericArray as unknown[]).length).toBe(5);
-    });
-
-    it('Then should support buffer-like operations', () => {
-      // Arrange
-      const operationalBuffer: _TTestBuffer = {
-        items: [],
-        count: 0,
-        maxSize: 100,
-      };
-
-      // Act - Simulate buffer operations
-      const items = operationalBuffer.items as unknown[];
-      items.push('item1', 'item2', 'item3');
-      operationalBuffer.count = items.length;
-
-      // Assert
-      expect((operationalBuffer.items as unknown[]).length).toBe(3);
-      expect(operationalBuffer.count).toBe(3);
-      expect(operationalBuffer.maxSize).toBe(100);
-    });
-
-    it('Then should handle test scenario data structures', () => {
-      // Arrange - Common test scenarios
-      const testScenarioBuffer: _TTestBuffer = {
-        testName: 'validation-test',
-        input: { param1: 'value1', param2: 42 },
-        expected: { result: 'success', code: 200 },
-        actual: null, // To be filled during test execution
-        status: 'pending',
-      };
-
-      // Act & Assert - Verify test structure
-      expect(testScenarioBuffer.testName).toBe('validation-test');
-      expect(testScenarioBuffer.input).toBeDefined();
-      expect(testScenarioBuffer.expected).toBeDefined();
-      expect(testScenarioBuffer.actual).toBeNull();
-      expect(testScenarioBuffer.status).toBe('pending');
-    });
-
-    it('Then should maintain type flexibility for testing needs', () => {
-      // Arrange
-      const flexibleBuffer: _TTestBuffer = {};
-
-      // Act - Add various types of test data
-      flexibleBuffer.mockData = { user: 'test', id: 123 };
-      flexibleBuffer.errorStates = ['ERROR', 'FATAL', 'WARNING'];
-      flexibleBuffer.validators = [(x: unknown) => x !== null, (x: unknown) => typeof x === 'string'];
-      flexibleBuffer.metadata = new Map([['key1', 'value1'], ['key2', 'value2']]);
-
-      // Assert - All types should be accepted
-      expect(flexibleBuffer.mockData).toBeDefined();
-      expect(Array.isArray(flexibleBuffer.errorStates)).toBe(true);
-      expect(Array.isArray(flexibleBuffer.validators)).toBe(true);
-      expect(flexibleBuffer.metadata instanceof Map).toBe(true);
-    });
-  });
-});
