+++ shared/packages/types/src/__tests__/unit/ErrorSeverity.spec.ts
@@ -1,4 +1,4 @@
-// src: src/__tests__/ErrorSeverity.spec.ts
+// src: src/__tests__/unit/ErrorSeverity.spec.ts
 // @(#) : Unit tests for ErrorSeverity enum and validation functions
 //
 // Copyright (c) 2025 atsushifx <http://github.com/atsushifx>
@@ -10,65 +10,14 @@
 import { describe, expect, it } from 'vitest';
 
 // Type definitions and utilities
-import { ErrorSeverity, isValidErrorSeverity } from '../../types/ErrorSeverity.types.js';
+import { isValidErrorSeverity } from '../../../types/ErrorSeverity.types.js';
 
 /**
  * Test suite for ErrorSeverity enum constant values.
  * Verifies that enum members have correct string values.
  */
 describe('Given ErrorSeverity enum values', () => {
-  describe('When accessing enum constants', () => {
-    it('Then should return fatal string value', () => {
-      // Act & Assert
-      expect(ErrorSeverity.FATAL).toBe('fatal');
-      expect(ErrorSeverity.ERROR).toBe('error');
-      expect(ErrorSeverity.WARNING).toBe('warning');
-      expect(ErrorSeverity.INFO).toBe('info');
-    });
-  });
-
   describe('When validating with isValidErrorSeverity', () => {
-    it('Then 正常系：should return true for all valid severity values', () => {
-      // Arrange
-      const validValues = [
-        'fatal',
-        'error',
-        'warning',
-        'info',
-        ErrorSeverity.FATAL,
-        ErrorSeverity.ERROR,
-        ErrorSeverity.WARNING,
-        ErrorSeverity.INFO,
-      ];
-
-      // Act & Assert
-      validValues.forEach((value) => {
-        expect(isValidErrorSeverity(value)).toBe(true);
-      });
-    });
-
-    it('Then 異常系：should return false for invalid types', () => {
-      // Arrange
-      const invalidValues = [
-        'invalid',
-        123,
-        null,
-        undefined,
-        true,
-        false,
-        {},
-        { severity: 'error' },
-        [],
-        ['error'],
-        () => 'error',
-      ];
-
-      // Act & Assert
-      invalidValues.forEach((value) => {
-        expect(isValidErrorSeverity(value)).toBe(false);
-      });
-    });
-
     it('Then エッジケース：should handle edge case values', () => {
       // Arrange
       const edgeCaseValues = [
